requirements-dev.txt
# This file was autogenerated by uv via the following command:
#    uv export --format requirements-txt --extra dev --no-hashes
-e .
annotated-types==0.7.0
    # via pydantic
backports-tarfile==1.2.0 ; python_full_version < '3.12' and platform_machine != 'ppc64le' and platform_machine != 's390x'
    # via jaraco-context
black==25.1.0
    # via superquantx
build==1.3.0
    # via superquantx
certifi==2025.8.3
    # via requests
cffi==1.17.1 ; platform_machine != 'ppc64le' and platform_machine != 's390x' and platform_python_implementation != 'PyPy' and sys_platform == 'linux'
    # via cryptography
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via
    #   black
    #   typer
colorama==0.4.6 ; os_name == 'nt' or sys_platform == 'win32'
    # via
    #   build
    #   click
    #   loguru
    #   pytest
contourpy==1.3.2 ; python_full_version < '3.11'
    # via matplotlib
contourpy==1.3.3 ; python_full_version >= '3.11'
    # via matplotlib
coverage==7.10.6
    # via
    #   pytest-cov
    #   superquantx
cryptography==45.0.7 ; platform_machine != 'ppc64le' and platform_machine != 's390x' and sys_platform == 'linux'
    # via secretstorage
cycler==0.12.1
    # via matplotlib
distlib==0.4.0
    # via virtualenv
docutils==0.21.2
    # via readme-renderer
exceptiongroup==1.3.0 ; python_full_version < '3.11'
    # via pytest
execnet==2.1.1
    # via pytest-xdist
filelock==3.19.1
    # via virtualenv
fonttools==4.59.2
    # via matplotlib
id==1.5.0
    # via twine
identify==2.6.14
    # via pre-commit
idna==3.10
    # via requests
importlib-metadata==8.7.0 ; (python_full_version < '3.10.2' and platform_machine == 'ppc64le') or (python_full_version < '3.10.2' and platform_machine == 's390x') or (python_full_version < '3.12' and platform_machine != 'ppc64le' and platform_machine != 's390x')
    # via
    #   build
    #   keyring
iniconfig==2.1.0
    # via pytest
jaraco-classes==3.4.0 ; platform_machine != 'ppc64le' and platform_machine != 's390x'
    # via keyring
jaraco-context==6.0.1 ; platform_machine != 'ppc64le' and platform_machine != 's390x'
    # via keyring
jaraco-functools==4.3.0 ; platform_machine != 'ppc64le' and platform_machine != 's390x'
    # via keyring
jeepney==0.9.0 ; platform_machine != 'ppc64le' and platform_machine != 's390x' and sys_platform == 'linux'
    # via
    #   keyring
    #   secretstorage
joblib==1.5.2
    # via scikit-learn
keyring==25.6.0 ; platform_machine != 'ppc64le' and platform_machine != 's390x'
    # via twine
kiwisolver==1.4.9
    # via matplotlib
loguru==0.7.3
    # via superquantx
markdown-it-py==3.0.0
    # via rich
matplotlib==3.10.6
    # via superquantx
mdurl==0.1.2
    # via markdown-it-py
more-itertools==10.8.0 ; platform_machine != 'ppc64le' and platform_machine != 's390x'
    # via
    #   jaraco-classes
    #   jaraco-functools
mypy==1.17.1
    # via superquantx
mypy-extensions==1.1.0
    # via
    #   black
    #   mypy
networkx==3.4.2 ; python_full_version < '3.11'
    # via superquantx
networkx==3.5 ; python_full_version >= '3.11'
    # via superquantx
nh3==0.3.0
    # via readme-renderer
nodeenv==1.9.1
    # via pre-commit
numpy==1.26.4
    # via
    #   contourpy
    #   matplotlib
    #   scikit-learn
    #   scipy
    #   superquantx
packaging==25.0
    # via
    #   black
    #   build
    #   matplotlib
    #   pytest
    #   twine
pathspec==0.12.1
    # via
    #   black
    #   mypy
pillow==11.3.0
    # via matplotlib
platformdirs==4.4.0
    # via
    #   black
    #   virtualenv
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
pre-commit==4.3.0
    # via superquantx
py-cpuinfo==9.0.0
    # via pytest-benchmark
pycparser==2.22 ; platform_machine != 'ppc64le' and platform_machine != 's390x' and platform_python_implementation != 'PyPy' and sys_platform == 'linux'
    # via cffi
pydantic==2.11.7
    # via superquantx
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.2
    # via
    #   pytest
    #   readme-renderer
    #   rich
pyparsing==3.2.3
    # via matplotlib
pyproject-hooks==1.2.0
    # via build
pytest==8.4.2
    # via
    #   pytest-benchmark
    #   pytest-cov
    #   pytest-mock
    #   pytest-xdist
    #   superquantx
pytest-benchmark==5.1.0
    # via superquantx
pytest-cov==6.3.0
    # via superquantx
pytest-mock==3.15.0
    # via superquantx
pytest-xdist==3.8.0
    # via superquantx
python-dateutil==2.9.0.post0
    # via matplotlib
pywin32-ctypes==0.2.3 ; platform_machine != 'ppc64le' and platform_machine != 's390x' and sys_platform == 'win32'
    # via keyring
pyyaml==6.0.2
    # via pre-commit
readme-renderer==44.0
    # via twine
requests==2.32.5
    # via
    #   id
    #   requests-toolbelt
    #   twine
requests-toolbelt==1.0.0
    # via twine
rfc3986==2.0.0
    # via twine
rich==14.1.0
    # via
    #   superquantx
    #   twine
    #   typer
ruff==0.12.12
    # via superquantx
scikit-learn==1.7.1
    # via superquantx
scipy==1.15.3 ; python_full_version < '3.11'
    # via
    #   scikit-learn
    #   superquantx
scipy==1.16.1 ; python_full_version >= '3.11'
    # via
    #   scikit-learn
    #   superquantx
secretstorage==3.3.3 ; platform_machine != 'ppc64le' and platform_machine != 's390x' and sys_platform == 'linux'
    # via keyring
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
threadpoolctl==3.6.0
    # via scikit-learn
tomli==2.2.1 ; python_full_version <= '3.11'
    # via
    #   black
    #   build
    #   coverage
    #   mypy
    #   pytest
twine==6.2.0
    # via superquantx
typer==0.17.4
    # via superquantx
typing-extensions==4.15.0
    # via
    #   black
    #   exceptiongroup
    #   mypy
    #   pydantic
    #   pydantic-core
    #   superquantx
    #   typer
    #   typing-inspection
    #   virtualenv
typing-inspection==0.4.1
    # via pydantic
urllib3==2.5.0
    # via
    #   requests
    #   twine
virtualenv==20.34.0
    # via pre-commit
win32-setctime==1.2.0 ; sys_platform == 'win32'
    # via loguru
zipp==3.23.0 ; (python_full_version < '3.10.2' and platform_machine == 'ppc64le') or (python_full_version < '3.10.2' and platform_machine == 's390x') or (python_full_version < '3.12' and platform_machine != 'ppc64le' and platform_machine != 's390x')
    # via importlib-metadata
